Provide examples of different types of RedactableArguments
Motivation
==========
Consulting the log redaction PRD to determine whether, for example,
a bucket name is "system" data or "meta" data is time-consuming.
It would be nice if that information were closer to hand.

Modifications
=============
Copy the data classification examples from the PRD into the Javadoc
for RedactableArgument.user(), meta(), and system().

Definitions come from this requirements document:
https://docs.google.com/document/d/1rbQsiDpKspJpzZspAMxjvU19cHEtvEFeGQ_jLmFJ4bM/edit#heading=h.q6wzxxq9l7kg

Result
======
Easier for developers to consistently and correctly classify
the different kinds of redactable data.

Change-Id: I766b3ae025a6b1675dd4a477646d7172173a28ab
Reviewed-on: http://review.couchbase.org/86483
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Build Bot <build@couchbase.com>