JCBC-1147: Add health check ping command support.
Motivation
----------
In addition to the refactored diagnostics command,
the RFC calls for a bucket-level "service ping"
facility which has similar output to the diagnostics
but proactively calls the services on their endpoints.

Modifications
-------------
This, fairly large (sorry) changeset adds "ping"
requests and responses to each service type, modifies
the endpooint handling code to deal with it and provides
the overall "HealthPinger" utility class which does
all the heavy lifting and aggregation.

Similar to the diagnostics API, it provides a JSON
export as well as experimental API access to the
actual objects.

Result
------
core work for health check ping support done, on
top of this in java-client the actual API binding
will be provided.

Change-Id: I0175e387ee2d59b77347cd74e677da17f74c4b21
Reviewed-on: http://review.couchbase.org/86658
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>