JVMCBC-484: Add warning on resource management to env javadoc.
Motivation
----------
It has been reported that while the docs point out that related
resources (buckets, clusters) need to be shut down before the
environment is, it should also be called out for in the
javadoc.

Modifications
-------------
The shutdown sync and async methods now contain a block which
state this fact.

Result
------
More clarity around resource management and shutdown of an
environment.

Change-Id: I0cc6eb13726c9ac896ddce2a58a8a90f9af7ce19
Reviewed-on: http://review.couchbase.org/87004
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Carabine <matt.carabine@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>