JVMCBC-485: don't go temporarily below min pool size
Motivation
----------
The code as it worked before closed all idle sockets and then
had some code in place which would "refill" the sockets up to
the point where its at min level again.

While this is fine, we need to make sure to stay at least at
the minimum level, even if it means that some sockets are idle
longer than configured.

Modifications
-------------
The code change leaves everything in place, but adds another check
where a maximum removable quota is defined before iterating. If
the quota is reached, no further endpoints are to be removed even
if they are idle.

Since the pool is likely to be dynamic, those will be recycled
if not needed anymore later on.

There is also a bit of a test refactoring to reduce the explicit
3s sleeps down to a linear backoff, leading to a 50% reduction
of the test execution time.

Result
------
The minimum quota is respected on the idle timer, and also as a
result the service never goes into a DISCONNECTED state since
it will never reach 0.

Change-Id: I4203c935ee6da7081ee140a5b2668fcda0eca625
Reviewed-on: http://review.couchbase.org/87005
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>