JVMCBC-486: Do not log full stacks on reconnect attempts.
Motivation
----------
This is a small change, but it greatly helps with visibility
in the logs when reconnects happen.

Modifications
-------------
For the majority of logs printed during reconnect attempts,
the logger now only logs the exception type and message (if
exists).

The stack trace is not helpful anyways since it comes somewhere
out of the netty async context and the message must be enough
to figure out whats going on. It also helpy to greatly reduce the
noise in the logs if such reconnects happen since it is "spammed"
with stacks that are not useful information.

Result
------
Less noisy in the logs when reconnect attempts happen.

Change-Id: I5f7c235284d37641445eca6bdef6478565ab8f20
Reviewed-on: http://review.couchbase.org/87788
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>