JVMCBC-480: Infrastructure for redacted logging (only redact user data)
Motivation
==========
The exact syntax for "system" and "meta" redaction is yet to be determined.
In the mean time, we've been asked to redact *only* "user" data.

Modifications
=============
Skip redaction for "system" and "meta" arguments.

Result
======
Only "user" data is redacted. Client code can still mark arguments
as "meta" or "system" -- those may be redacted in the future.

Change-Id: I0473b9e1c8aa185abb32fa72e4ed82304cdf67ce
Reviewed-on: http://review.couchbase.org/87871
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>