JVMCBC-487: fix netty epoll.so renaming
Motivation
----------
Couple versions ago our shaded dependency netty changed the
naming of the native epoll transport .so file to something
different, and we didn't track the rename change proprely.

As a result, the epoll transport couldn't be used with
recent SDK versions and failed during native lib load.

Modifications
-------------
The change is rather simple - it fixes the renaming
in the shaded jar file (as part of the pom process)
so that the proper one is included in our jar and
picked up as needed.

Result
------
the native epoll transport now works again out of
the box.

Change-Id: Ic0d8f4d40865ce6479869384ea8ed46e90901b88
Reviewed-on: http://review.couchbase.org/88018
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>