JVMCBC-489: Start node round-robin from offset.
Motivation
----------
Right now each query starts from index 0, which means it targets
the first node in the config list first. For short-lived queries
or many app servers that is not a problem, but it can become
one if many "fresh" apps hit n1ql queries that are long running
more or less at the same time.

Modifications
-------------
To avoid this situation, the code now initializes the counter
with a value between 0 and 1023 (incl) to give it some initial
"randomness". While it does not remove the possibility especially
with a low number of query (or other service) nodes to hit the same
one, it reduces the likelyhood.

Since the randomness made the test cases nondeterministic, helper
constructors have been used which allow the tests to initialize
the counter to 0 so that they work reliably.

While being in this code area, I notice that the volatile counter
increment wasn't atomic. While its not really an issue since we
do not rely on atomic increments, I changed it to a AtomicLong
so the code is now correct in this regard.

Finally, I noticed that the analytics locator didn't have tests
like the others, so I added one.

Result
------
Round-robin based node locators (all but kv) are now starting
from a random node so they not all hit the same node the first
time on startup.

Change-Id: I05db854242c80b70ef1aa7790eabb8b63bf1ddc7
Reviewed-on: http://review.couchbase.org/88179
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>