JVMCBC-491: Create CoreContext.
Motivation
----------
This is the first of a couple changesets which propagate
a CoreContext down into the individual components
rather than individual objects independently.

This makes it a little bit easier to reason about core-global
state, but also allows us in the future to add more
context (for example the core-identifier we need for the new
HELLO).

Modifications
-------------
This changeset adds the initial context and propagates it down
to the first layer, the RequestHandler.

Result
------
CoreContext created and first layer propagation.

Change-Id: Iff86d0ed5e1d1ae99b176dfda6eb8c5413541572
Reviewed-on: http://review.couchbase.org/89002
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>