JVMCBC-492: Create unique ID per core
Motivation
----------
This changeset introduces a unique ID per core instance and
propagates it into the context so it can be used in
the endpoints.

Modifications
-------------
In addition to implementing it in the core itself, the feature
handler now also accepts the context. In a future changeset,
the handler will then use this ID for the hello
negotiation.

Result
------
A unique ID per core and propagated.

Change-Id: Ia976e2a8008215e167216fda19c41aa22d8ad520
Reviewed-on: http://review.couchbase.org/89009
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>