JVMCBC-493: Negotiate JSON user agent in HELLO
Motivation
----------
As part of better tracing capabilities, the client must
negotiate an identifer in addition to the user agent.

Based on the RFC this format is a JSON hash with both
the agent and the identifier.

Modifications
-------------
This changeset uses the new CoreContext and its core ID
to negotiate the HELLO with a JSON blob. It also uses the
channel ID as a transient identifier as outlined in the RFC.

Simple unit tests in place which especially make sure that
the long values are hex padded as well as too long user
agent strings are cut off at 200 chars as per RFC.

Result
------
Properly negotiated JSON-based user agent in HELLO.

Change-Id: I848a8dd986b0a50ca47315f0ad99eea3354e9e4e
Reviewed-on: http://review.couchbase.org/89011
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>