JVMCBC-495: Make CoreEnvironment.Builder generic
Motivation
----------
Downstream implementations of the environment builder right now
need to override each builder method to keep the fluent API going.
This leads to redundant code as well as increases the error rate
if we forget to add the downstream builder overrides.

Modifications
-------------
This changeset turns the CoreEnvironment.Builder into a generic
builder based on the common "generic builder pattern" java
idiom.

The result of this is that downstream implementations can
basically throw away all their overrides and going forward we
don't need to add more overrides as we progress.

Result
------
Less code duplication and less error rate during development
of new features thanks to the generic builder pattern.

Change-Id: I88f746496b15cb363ca02fbe9379c66aeaf8235d
Reviewed-on: http://review.couchbase.org/89234
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>