JVMCBC-496: Add support for kv service tracing.
Motivation
----------
As part of the overall tracing efforts, the server side
(in this case, particularly the kv engine) adds the
capability to return its overall processing duration
as part of the response if negotiated.

Modifications
-------------
This changeset makes use of the new functionality and
does so in the following form:

 - It negotiates the tracing flag via HELLO
 - Extends the memcached netty protocol handlers to support
   the new 0x18 response magic including the new extended
   framing capabilities
 - Modifies the kv handler to parse the tracing data out
   of the frame if present and pass it to the BinaryResponse.

Result
------
It is now possible to negotiate and extract server side
duration when used via KV (for now).

Note that this functionality is going to be wired up in the
wider tracing context in future commits.

Change-Id: Ia4022447464202783c6ad58c415f0d84b10a35d7
Reviewed-on: http://review.couchbase.org/89252
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>