JVMCBC-497: Allow snappy only on certain ops.
Motivation
----------
Previously the code allowed to enable snappy on all operations
which have a memcached body payload. This turns out not to be
supported on the server side, so adjustments need to be made.

Modifications
-------------
This changeset limits the compressible operations to the list of
server supported ones: insert, upsert, replace, append and
prepend.

Result
------
When compression is enabled, only the supported operations
are actually compressed.

Change-Id: I0c70e5ba30060bef09f243b5fd3090a754e7bdba
Reviewed-on: http://review.couchbase.org/89256
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>