JVMCBC-498: Add OpenTracing infrastructure.
Motivation
----------
This changeset expands the opentracing infrastructure to include
it as a dependency and hooks up the environmental
accessor accordingly.

The way it is set up right now enables the upcoming slow log
tracer by default if opentracing is enabled. If manually
disabled, the bundled Noop tracer will be used instead.

Change-Id: I4eb2773b2f69d70bcec2e461f3023d7280f20968
Reviewed-on: http://review.couchbase.org/89314
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>