Gardening: cleanup false positive leaks.
Motivation
----------
By default we've been logging our tests at DEBUG level which
adds a lot of noise to the output and makes important pieces
vanish.

Modifications
-------------
This changeset first and foremost sets the default logging
to WARN so warnings and stack traces are much easier to
spot. As a result, the following insights have been gained
and resolved:

 - Serveral KV unit tests were lacking the required buffer
   release and in turn lead to false-positive leak
   warnings.
 - The refresher sometimes showed NPEs which were not harmful
   but came from the fact that the provider is set after
   init but the polling starts quickly. So the code now
   makes sure its picked up via volatile but also checks
   if its not null before trying to access it.
 - a stray println was discovered and subsequently removed.

NOTE: there has been one legit issue discovered with
the search handler and leaking buffer on ping and
keepalive, but this is enough of a real bug to
warrant a separate changeset and a real ticket.

Result
------
Its now easier to spot issues and some of them got fixed.

Change-Id: I4b929940a6330f5055c5a30fa70b6e178c90ef43
Reviewed-on: http://review.couchbase.org/89400
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>