JVMCBC-500: Bump dependencies to latest bugfix releases.
Motivation
----------
As part of the regular bugfix dep bump, this change makes sure
all of our dependencies are up-to-date.

Modifications
-------------
Note that we previously would pass "null" into the channel promise
but in recent version a check has been put in place. Netty privately
packages a similar FailedChannel, but since its not publicly
accessible and a simple class this changeset just copies it over.

Result
------
Dependencies bumped.

Change-Id: I083e7e52cb6bbcb230d9af8ad90819bb90342b78
Reviewed-on: http://review.couchbase.org/89405
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>