JVMCBC-498: Add ThresholdLogTracer
Motivation
----------
As part of the tracing infrastructure, the client need to
ship out of the bux with a tracer which supports collecting
information about spans that are over the treshold.

Modifications
-------------
The ThresholdLogTracer and its companion classes provide
this kind of functionality, they just represent in memory
objects of this span kind.

Note that for simplicity reasons some functionality has
been omitted, like injecting or extracting as well as
span event logging because they are not needed for this
tracer and would just add overhead.

Note that the actual logging via a reporter is done in
the upcoming changeset.

Result
------
Support for the ThresholdLogTracer with its related
class implementations.

Change-Id: Id562e44505e092129c1626db8a24ad6a42f28152
Reviewed-on: http://review.couchbase.org/89476
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>