JVMCBC-498: Add ThresholdLogReporter.
Motivation
----------
Building on the previous changesets, this
one adds a reporter and provides the
infrastructure and constructs for logging.

Modifications
-------------
The reporter has been added as well as a very basic
test added. Future commits will flesh out the
span collection and logging at the same time.

Result
------
Threshold reporting infrastructure in place.

Change-Id: Ibe004abbaa5dc7a983fc0ba731c6c4a8c7241b2b
Reviewed-on: http://review.couchbase.org/89558
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>