JVMCBC-503: Switch to different snappy for compression.
Motivation
----------
Originally we've been using the snappy code from netty,
but since we found issues in both the encoder and
decoder side the stability of it has been in question.

Modifications
-------------
This changeset includes the new dependency, shades it
and uses it in the KV handler for snappy compression
and decompression.

Result
------
Correctly working snappy implementation, more confidence
in that it works properly in production.

Change-Id: I85670b3bfe7ae893920fe7ff3c45b16fbea99f66
Reviewed-on: http://review.couchbase.org/89657
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>