JVMCBC-498: Implement Span Tracing
Motivation
----------
Building on the earlier changes, this changeset is intended
to actually wire up the span tracing with the required
attributes.

Modifications
-------------
All request types and the handlers are extended to properly
log and enrich the spans (including the dispatch span)
as well as adding additional information to the parent
span where possible.

Result
------
core-level span tracing and sending into the reporter.

Change-Id: I188969587dfd5e7dd04423624d5357e0111d9dbf
Reviewed-on: http://review.couchbase.org/89864
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>