JVMCBC-507: Make sure core id is random and long.
Motivation
----------
Previously the core id would start at 0 all the time (1 as the
first core), but this makes it not unique across multiple
application servers.

Also, it should actually be stored as a long since its expected
to be dumped as 16 char hex value.

Modifications
-------------
To increase the entropy of the core id, it is initialized at
a random integer. Note that if its below 0, we force it positive
so that its also in bounds.

Result
------
When many app servers are used, its now easier to distinguish
them in the logs and via tracing.

Change-Id: I28e09901b3187dfaf32b1876a3337e8b204ddc9f
Reviewed-on: http://review.couchbase.org/90131
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>