JVMCBC-510: Ignore unknown BucketCapabilities.
Motivation
----------
It was found that with the default settings, jackson
is pretty strict about what happens if a value cannot
be matched to an enum value, causing problems when
the server adds a new capability that the client
does not understand.

Modifications
-------------
This changeset aims to remedy this situation by setting
a jackson deserialization feature which turns them into
"null", enough for our uses cases and the least intrusive
option.

A test case has been added to make sure the config parses
without issues with an unknown bucket capability. Note that
while in there, I found that the bucket config test cases
should actually go through the BucketConfigParser in order
to test the identical path the actual code also goes through
instead of using jackson directly. Better decoupling as
well as making it hits the right codepath.

Result
------
The server can now add features without the client needed
to be patched immediately so that it doesn't break.

Change-Id: Idce7b96ab1de19b5a91b225c933f97142fdbd5eb
Reviewed-on: http://review.couchbase.org/90441
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>