JVMCBC-513 Handle NPE on a bad host in connection string
Motivation
----------
The client will exit with a NPE if a connection string contains both
good and bad hosts.

Changes
-------
Connection string adds a static method that will check if the
hosts are resolvable. The older static method parseHosts and a list of
unresolved hosts is retained for unit tests.

Results
-------
The bootstrap continues successfully with bad hosts. The integration
test will be added in the client.

Change-Id: I2889f0d80fa7b4aecdf613143c45eadef0d58722
Reviewed-on: http://review.couchbase.org/90566
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>