JVMCBC-521: Wrap raw config into context.
Motivation
----------
As a first step of fixing JVMCBC-521, the code needs to be prepared
to send around not the raw config alone but more information
associated.

Modifications
-------------
A new ProposedBucketConfigContext is added as a value object which
contains the bucket name and its origin apart from the actual raw
config.

The rest of the changes is just passing around the value object
instead of the raw config and adapting tests for the new
signature.

Result
------
More flexibility when passing around a raw, proposed config.

Change-Id: I10c5a3978fc57db0cd5a460539486d3fcdfbe4cf
Reviewed-on: http://review.couchbase.org/92512
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>