JVMCBC-521: Use origin instead of localhost for nodesExt
Motivation
----------
If the hostname from nodesExt is missing, the code should not
use the hardcoded localhost but rather the origin (similar to
the replace ops already in place).

Modifications
-------------
This changeset propagates the origin from the outer layers through
the json config deserialization and then applies it to the abstract
bucket decoding.

Result
------
The proper origin hostname is used and not the hardcoded localhost.

Change-Id: If8547fc69f911da00784b98bbcff29c8d4b09749
Reviewed-on: http://review.couchbase.org/96149
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>