JVMCBC-525: Make sure compression only reads allowed slice
Motivation
----------
During bugfixing it was found that the compression code is only
allowed to read the readable bytes into an array and not just look
at the full array 1:1, especially if during buffer pooling other
chunks might be filled with other stuff.

Modifications
-------------
Write the readable bytes into a new array and use this for compression,
which cleans up subdoc issues as well as other test failures.

Result
------
Proper compression bytes are sent over the wire now.

Change-Id: I00948ede2b9064c365a8cf9d2d8cae42fc2b531c
Reviewed-on: http://review.couchbase.org/92709
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>