JVMCBC-528: Provide more information during sasl failure.
Motivation
----------
As mentioned in the ticket, for whatever reason the server might
during the challenge phase not return the expected format but
something else like a JSON string with an error message.

This is technically not something we can fix at this point, but
we can improve the debug experience around it by properly logging
what the server returns.

Modifications
-------------
The challenge code is modified so that it catches attribute decoding
exceptions and carries them forward with a descriptive error message.

Result
------
Better debugging experience during the SASL challenge phase.

Change-Id: Idd2b9f65dc218f05038ca27e59c0786be266aefa
Reviewed-on: http://review.couchbase.org/92844
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>