JVMCBC-535: SDK should validate final SASL response
Motivation
----------
SDK needs to authenticate the server by computing the ServerSignature and comparing it to the value sent by the server

Modifications
-------------
This changeset modifies code evaluating the final response sent by the server.

Result
------
Verified with CouchbaseMock and wireshark. If you remove the serversignature sent by the final message:
* Before: the client consider the authentication exchange is successul => KO
* After: the client consider the authentication exchange is unsuccessul => OK

Change-Id: I0056fa184ec24dc09e53ba2b9d7051c877def03e
Reviewed-on: http://review.couchbase.org/92962
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>