JVMCBC-529: Expose original mutation cas on observe exceptions.
Motivation
----------
When running a mutation with PersistTo/ReplicateTo and the actual
mutation succeeds but the observe fails, it makes sense to expose
the original CAS value so it can be used for further processing.

Modifications
-------------
This changeset slightly extends the current exceptions raised
during the observe phase and extends them with the cas value
from the original mutation. Since the exception can be thrown
in non-mutation cases (read from replica) the cas will be
0 in this case - the javadoc mentions that fact.

Result
------
Exceptions during the observe phase expose the cas value
where possible from the original mutation.

Change-Id: I30929de0cf84efeeaa1dae7a307aa89d6d9ba7fa
Reviewed-on: http://review.couchbase.org/93510
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>