JVMCBC-523: Properly clean up services on failed connect.
Motivation
----------
It was found that when the bootstrap list node is mixed with
kv and non-kv nodes, the non-kv nodes won't be able to bootstrap.
Right now it all works, but its not cleaning up the sockets
(services) behind that can't be used - this leads to exceptions
in the logs and in general not the behavior intended.

Modifications
-------------
This changeset modifies code in a couple of places, basically
making sure that when during bootstrap the loader fails its
actively tearing down the service. As part of implementing this,
I found that the service didn't properly disconnect and a potential
race condition with the disconnected states came up so this has
been also cleaned up.

One test case has been removed *sadpanda* since it tested the
exact same thing I had to remove. Note that this doesn't change
anything since the underlying endpoint doesn't care and
performs the same check anyways.

NOTE: You will still see another round of Node Connected
and Node Disconnected attempts. This is because the loader
chain is set up in a way that HTTP is tried next but it will
fail again obviously - this won't do any harm but cleaning
that up can be done in another commit and might be a little
more complex based on initial investigation.

Result
------
Proper close of services/sockets on those nodes where bootstrap
couldn't complete.

Change-Id: I50a7d0a17fcd786663df62b0c5f233227111d624
Reviewed-on: http://review.couchbase.org/93414
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Reviewed-on: http://review.couchbase.org/93503
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>