JVMCBC-534: PooledService creates excessive endpoints on sending to downed node
Motivation
----------
If Couchbase server is stopped, and a N1QL query is then made to it,
an excessive number of endpoints is created - well above the maximum
configured.

This happens because PooledService will create an endpoint, it will
timeout after 32 msecs, and this lead to 'unsubscribeAndRetry' being
called and the request being tried again, leading to another endpoint
being created.  But, the original endpoint is still around.  This loops
repeatedly until a 300s timeout on the original endpoint fires, killing
it.

There is a mechanism that's meant to prevent this, 'pendingRequests',
and it's correctly incremented in 'maybeOpenAndSend', but then
32msecs later it's decremented in 'unsubscribeAndRetry'.  So
pendingRequests is now 0 but the created endpoint remains extant.

Modifications
-------------
Remove 'pendingRequests'.

On creating an endpoint, add it directly to the 'endpoints' variable.

Modified existing unit tests and added another.

Unrelated: removed sonatype plugin as it refuses to work on my laptop.
This was discussed in another (pending) review.

Result
------
The endpoint will still timeout after 32msecs, and 'unsubscribeAndRetry'
will still try to resend the request.  But now the 'send' logic won't
allow more endpoints than its configured maximum to be created (max 12).

I've confirmed that all the 'selectionStrategy' options handle endpoints
that aren't CONNECTED yet.

All tests pass.  In addition performed manual testing to confirm e.g.
once server is restored that subsequent N1QL requests function.

Change-Id: I051b93ab4bf0ae72e17dd1c6f12428b62c40ac6a
Reviewed-on: http://review.couchbase.org/93868
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-on: http://review.couchbase.org/95133
Tested-by: Michael Nitschinger <michael@nitschinger.at>