JVMCBC-535: Make sure PLAIN is bailed out early.
Motivation
----------
The previous changeset left out one possible complication, namely
that the PLAIN auth method finished earlier and so the evaluate
challenge step failed.

Modifications
-------------
This changeset makes sure that before the challenge is evaluated,
it is checked if the exchange already finished. Also in the process
it was found that the response might already come back with an
auth error as part of the status, so in this case we can short-circuit
and fail the process right away.

Result
------
Proper handling of PLAIN SASL auth as well as short-circuiting sasl
auth response status codes.

Change-Id: I37e718d7e2a252c91ba7e1747378a9da993ddbf3
Reviewed-on: http://review.couchbase.org/94201
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Senol Ozer <senol.ozer@amadeus.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>