JVMCBC-537: don't send http auth header if not needed
Motivation
----------
When using cert auth there is no username or password
than can or should be negotiated via basic auth. Omit
setting the header in this case.

Modifications
-------------
Since cert auth is not propagated semantically down
to the IO layer, what is used to signal is that both
user and password are null (or empty as a check). Since
if we neither have those credentials it doesn't make
sense to add the header anyways, so just skip it.

This also fulfills the contract of not sending the
http auth header when cert auth is used.

Result
------
Don't set the http auth header if username and
password are not available.

Change-Id: I57d95b5937228d5f50439c6f90cce4906b0339ac
Reviewed-on: http://review.couchbase.org/94206
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>