JVMCBC-482: Force IPv4 property to false by default
Motivation
----------
When the new NetworkAddress class was developed, to
preserve backwards compatibility the Force IPv4
property was set to false in order to not
cause issues on dual-stack systems which relied
on the old behavior.

Modifications
-------------
Note that JVMCBC-475 outlined that with the next
minor bump the switch will be flipped off by
default which is what this changeset is doing.

Users can always restore the old behavior by
setting the system property manually, but
since newer server versions have proper IPv6
support the need for this should go away
over time.

Result
------
Proper IPv6 support out of the box in dual
stack systems with the option to switch back
to the old behavior if needed.

Change-Id: I8d82cb3209f5c5bb763ff7840b9d12ccab428853
Reviewed-on: http://review.couchbase.org/94653
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>