JVMCBC-541: Add threshold / orphan counters
Motivation
----------
The count property when processing threshold and orphans is meant to
represent the total number of spans that were reported during a given
interval. However, the current implementation uses the current service
set size, which is limited to the the sample size.

Modifications
-------------
- add counters for each set service set
- increment counter when processing new span
- use counter when generating JSON
- reset counters when clearing sets

Result
------
Threshold and orphan counts are accurate and not restricted to the size
of the span set.

Change-Id: I1a5fd984024268eafa934e516f2d7ef4620713b7
Reviewed-on: http://review.couchbase.org/94768
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>