JVMCBC-543: Recover earlier than OS TCP timeout
Motivation
----------
It has been found that when a TCP connection is in a
half-open state the app-level keepalive may not function
as intended since operations are still in-flight but
will never complete. Usually this situation resolves
itself when the TCP idle timers/keepalives fire on the
OS level, although this can take a long time and is
not reliable enough since it depends on config params
which we do not control.

Modifications
-------------
This changeset helps to alleviate the situation by
making the following changes to how keepalives are
handled inside the IO layer:

 1) Keepalives are now also sent into the channel not
    only if there is no traffic, but also if all the
    operations stuck in the sent queue are not active
    anymore (speak: timed out).
 2) The write-path has some more special conditions to
    check if the keepalive is allowed to go through
    compared to a regular operation (which needs
    to be rescheduled if pipelining is enabled and
    something is in-flight already).

Note that at the upper layers (java-client) we need
to be careful to properly supply the subscriber
into the request otherwise the operation is deemed
active forever and this change won't work as expected.
Separate changesets handle this.

Result
------
Quickly detect half-open TCP connections by sending
the app level keepalives even if ops are in-flight
if they are timed out (non active) already.

Change-Id: I42cb0b95eb1af6f7ea0ced7ecf1c51f15227be54
Reviewed-on: http://review.couchbase.org/95043
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>