JVMCBC-552: Improve ring buffer diagnostics
Motivation
----------
When core-io's ringbuffer is filled, a BackpressureException is thrown
to the app.  This change is to provide some additional helpful
diagnostics in the exception.

Modifications
-------------
On adding a request to the ringbuffer, it's stored in the new
RingBufferMonitor class.  For lock-free performance, this class uses
a very simple approach of storing one AtomicInteger counter per
query type.

When requests are pulled from the ringbuffer, they are also removed
from the RingBufferMonitor.

If the ringbuffer ever fills, the BackpressureException raised now
contains additional diagnostics from RingBufferMonitor.

If this change is ok, I'll look at adding an API so that apps can query
the RingBufferMonitor at any point, and so monitor how close to
backpressure they are.

Special considerations for review:
- Is there a good place for the RingBufferMonitor to go, rather than
  having to be a singleton.
- I've had to not use the statically created BackpressureException,
  including some code that sets its stacktrace.  Unsure of the implications.
- Are the QueryTypes granular enough?
- Does it also need to be added to RetryHelper?  I think this is after the
  ringbuffer, but please confirm.

Result
------
Tested manually, as wasn't sure how to unit test this.  Counts are
increased and decreased as expected.

Change-Id: I03ba2f5e949c83b5c0b5441df901b7c414107429
Reviewed-on: http://review.couchbase.org/95183
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>