JVMCBC-548 Order not preserved when shuffling config provider seed hosts
Motivation
==========
In DefaultConfigProvider.seedHosts we shuffle the list of network
addresses, but then we put them back into a HashSet which does not
respect insertion order. In fact, the iteration order is often the same
after shuffling and inserting into the new HashSet as it was in the
original Set.

Modifications
=============
Store the seed hosts in a LinkedHashSet, which preserves iteration order.
For consistency, make a defensive copy even when not shuffling.

Result
======
Seed host iteration order is actually random.

Change-Id: Id30830d5ca2fd4a7c8ed7354ee7958f92acc4953
Reviewed-on: http://review.couchbase.org/95504
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>