Update mockito for testing.
Motivation
----------
When the tests are run with jdk 9+ the runtime complains about
illegal reflective access with our outdated version of mockito,
so this changeset upgrades it to the latest version.

Note: tests had to be modified because in mockito 1.x the anyString
matcher included null values while in 2.x it does not anymore.
the equivalent to this is nullable(String.class) which makes the
tests pass without changing its semantics.

Change-Id: I261a896faedf7b1bcdad616c9ae04871bf4f608e
Reviewed-on: http://review.couchbase.org/96144
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>