JVMCBC-542: Only use nodesExt to fetch alternates
Motivation
----------
Previously the nodes part of the config was used to
pick up alternate addr hostname and port, but nodesExt
needs to be used in order to also pick up SSL ports.

Modifications
-------------
This changeset adapts the PortInfo section (which represents
nodesExt) to also include the alternate address portion and
then modifies the config to set it directly from there rather
from the pure nodes section. The test case has been adapted
to also verify SSL ports are loaded since they are already
also part of this config.

Result
------
For alternate addresses, SSL ports are now also picked up
properly and the SDK in general looks in the right spot.

Change-Id: Ic518aedbc5216ab91ee8556adf58b875d692c94e
Reviewed-on: http://review.couchbase.org/96370
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>